Move qarray implementation to separate folder and split files
